Version 0.6. Input file paths can now be provided. All real*4 converted to real*8. i_le, t_le values initialized correctly.

+++ globvar.f90
@@ -34,7 +34,7 @@ integer cpdeltam,cpdeltheta,cpinbeta,cpoutbeta,cpchord,cptm_c
 integer, allocatable, dimension(:) :: ncp_curv,ncp_thk,throat_index,BGgrid_all
 integer n_inter_intervals,nsp_interpolated_hub
 !
-real*4, allocatable, dimension(:,:) :: bladedata,splinedata
+real*8, allocatable, dimension(:,:) :: bladedata,splinedata
 real*8 sinl,sext,thkc ,chrdx,mr1,stak_u,stak_v, x1hub,x1tip,r1hub,r1tip
 real*8 xb(nx),yb(nx),mp(nx,nax),xstk,xstk1,abs_zero
 real*8 xa(nx,nax),ya(nx,nax),xms(nx,nax),rms(nx,nax)
